requirements.txt
aiogram
python-dotenv
aiofiles
pytz